% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-FTSE100}
\alias{FTSE100}
\alias{data-FTSE100}
\title{Financial Times Stock Exchange 100 Index, 01.01.2015--04.11.2016}
\format{A bivariate time series with 468 observations; a \code{zoo} object}
\description{
Share index of 100 companies listed on the London Stock Exchange with the
highest market capitalisation. The first series contains the opening prices,
the second series contains the closing prices.
}
\details{
The data was downloaded from Finanzen.net on 4 November 2016.
}
\examples{
FTSE100_open <- FTSE100[,1]
FTSE100_close <- FTSE100[,2]
FTSE100_returns <- (FTSE100_open - FTSE100_close) / FTSE100_close
FTSE100_volatility <- FTSE100_returns^2 - mean(FTSE100_returns^2)
op <- par(mfcol=c(1,2))
plot(FTSE100_close, main = "", ylab="FTSE 100", xlab="year")
plot(FTSE100_volatility, main = "", ylab="returns (sq., ctrd.)", xlab="year")
par(op)
}
\references{
Obtained from Finanzen.net on 4 November 2016
\url{http://www.finanzen.net/index/FTSE_100/Historisch}

Further information on the index
\url{https://en.wikipedia.org/wiki/FTSE_100_Index}
}
\keyword{data}

