% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-Hohenpeissenberg}
\alias{data-Hohenpeissenberg}
\alias{Hohenpeissenberg}
\title{Daily air temperatures, collected in Hohenpeissenberg, 1985--2016}
\format{A univariate time series with 11680 observations; a \code{ts} object}
\description{
The station of Hohenpeissenberg is located in Germany, Brandenburg.
Its location is 977m (elevation), 47.8009 (latitude) and 11.0109 (longitude).
DWD lists it with station-id 2290. Data is avaiable from 1781 onwards.
}
\details{
The observations for any 29 February were removed to have 365 observations
per year.
}
\examples{
## Perform Fourier regression to remove trend and seasonality
library(forecast)
Y <- Hohenpeissenberg
tt = 1:length(Y)
year.lm = lm(Y ~ (tt + fourier(Y, 4)))
Y.mean <- ts(fitted(year.lm), f = 365, start=c(1985,1,1))
Y.norm = ts(Y - Y.mean, f = 365, start=c(1985,1,1))
op <- par(mfcol=c(1,2))
plot(Y, main = "", ylab="temperatures ", xlab="date")
lines(Y.mean, col="red")
plot(Y.norm, type="l", main = "", ylab="adjusted temperatures", xlab="date")
par(op)
}
\references{
Obtained from Deutscher Wetterdienst
\url{http://www.dwd.de/DE/klimaumwelt/cdc/cdc_node.html}
}
\keyword{data}
