% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastSNSTS-package.R
\docType{package}
\name{forecastSNSTSexamples-package}
\alias{forecastSNSTSexamples}
\alias{forecastSNSTSexamples-package}
\title{Data examples for the R package 'forecastSNSTS'}
\description{
Three demos that replicate the data analysis examples from Section 5 of
Kley et al (2016).
}
\details{
\tabular{ll}{
   \cr Package: \tab forecastSNSTSexamples
   \cr Type:    \tab Package
   \cr Version: \tab 1.0-0
   \cr Date:    \tab 2016-11-15
   \cr License: \tab GPL (>= 2) (for R code),
   \cr          \tab Open Government License (for data set)
 }
}
\author{
Tobias Kley
}
\references{
Kley, T., Preuss, P. & Fryzlewicz, P. (2016).
Predictive, finite-sample model choice for time series under stationarity and non-stationarity.
[cf. \url{https://arxiv.org/abs/1611.04460}]
}

